requirements.txt
gdown 
pandas
syllables
bs4
requests
pydrive
nltk
spacy
pyaml
lxml
openpyxl
